requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.3
mysqlclient==2.1.1
python-dotenv==1.0.0
WTForms==3.0.1
pytz